% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeWrapper.R
\name{initializeDataFrame}
\alias{initializeDataFrame}
\title{Create a Data Frame from pokeAPI call}
\usage{
initializeDataFrame()
}
\value{
A data frame
}
\description{
The initializeDataFrame() function sends two get requests to Pokémon API. The API calls
return data in the format of JSON and gets parsed by jsonlite. Data from both API
calls are appended to a data frame.
}
\examples{
\dontrun{
pokeframe <- initializeDataFrame()
}
}
