% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeWrapper.R
\name{poke.filter}
\alias{poke.filter}
\title{Create a Data Frame Filtered by Pokémon}
\usage{
poke.filter(pokeframe, vec)
}
\arguments{
\item{pokeframe}{The data frame of pokemon to summarise}

\item{vec}{the name(s) of the pokemon being filtered, accepts a string for a single pokemon, or a list for multiple}
}
\value{
A data frame
}
\description{
The poke.filter() function filters the pokeframe data frame by pokemon names
selected in a vector,
}
\examples{
\dontrun{
poke.filter(pokeframe, "jigglypuff")
poke.filter(pokeframe, c("pikachu", "bulbasaur", "charmander", "squirtle"))
}
}
