% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeWrapper.R
\name{poke.summary}
\alias{poke.summary}
\title{Create a Data Frame with Aggregated Data from pokeframe}
\usage{
poke.summary(pokeframe, infoType)
}
\arguments{
\item{pokeframe}{The data frame of pokemon to summarise}

\item{infoType}{The type of information to summarise 'habitat' or 'type'}
}
\value{
A data frame
}
\description{
The poke.summary() function counts the number of Pokémon and mean of capture rate grouped
by either 'habitat' or type' from pokeframe data frame.
}
\examples{
\dontrun{
poke.summary(pokeframe, 'habitat')
poke.summary(pokeframe, 'type')
}
}
