% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeWrapper.R
\name{poke_api}
\alias{poke_api}
\title{The poke_api() function takes in variable part of the path required
for the GET request to the pokeAPI servers.}
\usage{
poke_api(path)
}
\arguments{
\item{path}{The pokeAPI parameter to pull data for}
}
\value{
The json response from the pokeAPI
}
\description{
The poke_api() function takes in variable part of the path required
for the GET request to the pokeAPI servers.
}
\examples{
poke_api("/pokemon-species/1")
}
