% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeWrapper.R
\name{summary}
\alias{summary}
\title{Create a Data Frame with Aggregated Data from pokeframe}
\usage{
summary(pokeframe, infoType)
}
\arguments{
\item{pokeframe}{The data frame of pokemon to summarise}

\item{infoType}{The type of information to summarise 'habitat' or 'type'}
}
\value{
A data frame
}
\description{
The summary() function counts the number of pokemon and mean of capture rate grouped
by either 'habitat' or type' from pokeframe data frame.
}
\examples{
\dontrun{
summary(pokeframe, 'habitat')
summary(pokeframe, 'type')
}
}
